% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nn_classification_cpp}
\alias{nn_classification_cpp}
\title{Nearest Neighbor Classification}
\usage{
nn_classification_cpp(train, test, classes)
}
\arguments{
\item{train}{A numeric matrix of training points}

\item{test}{A numeric matrix of test points}

\item{classes}{A integer vector of training point classifications}
}
\description{
Finds the closest neighbor for each point in the test matrix
and returns a vector of classifications
}
